requirements.txt
altgraph==0.17
cycler==0.10.0
decorator==4.4.2
fbs==0.8.9
future==0.18.2
imageio==2.9.0
kiwisolver==1.2.0
macholib==1.14
matplotlib==3.3.0
natsort==7.0.1
networkx==2.4
numpy==1.19.1
pefile==2019.4.18
Pillow==7.2.0
PyInstaller==3.4
pyparsing==2.4.7
PySide2==5.15.0
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-video==1.1.11
scipy==1.5.2
shiboken2==5.15.0
six==1.15.0
tifffile==2020.7.24
